requirements.txt
# For pinned versions, see requirements.lock
anthropic>=0.40.0
mcp>=1.26.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.0.0
pyyaml
python-dotenv
pandas
plotly
kaleido
matplotlib
httpx
ddgs
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
python-jose[cryptography]>=3.3
bcrypt>=4.0
filelock>=3.12
alembic>=1.13
asyncpg>=0.29
redis[hiredis]>=5.0
arq>=0.26
structlog>=24.0
opentelemetry-api>=1.20
opentelemetry-sdk>=1.20
opentelemetry-instrumentation-fastapi>=0.41b0
prometheus-client>=0.20
boto3>=1.34
croniter>=2.0
psutil>=5.9
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
pytest-timeout>=2.3
